+++ src/_wrappers.c
@@ -15,6 +15,7 @@ typedef struct {
 
     PyObject *dict;
     PyObject *wrapped;
+    PyObject *weakreflist;
 } WraptObjectProxyObject;
 
 PyTypeObject WraptObjectProxy_Type;
@@ -48,6 +49,7 @@ static PyObject *WraptObjectProxy_new(PyTypeObject *type,
 
     self->dict = PyDict_New();
     self->wrapped = NULL;
+    self->weakreflist = NULL;
 
     return (PyObject *)self;
 }
@@ -152,6 +154,9 @@ static int WraptObjectProxy_clear(WraptObjectProxyObject *self)
 static void WraptObjectProxy_dealloc(WraptObjectProxyObject *self)
 {
     PyObject_GC_UnTrack(self);
+
+    if (self->weakreflist != NULL)
+        PyObject_ClearWeakRefs((PyObject *)self);
 
     WraptObjectProxy_clear(self);
 
@@ -1683,7 +1688,7 @@ PyTypeObject WraptObjectProxy_Type = {
     (traverseproc)WraptObjectProxy_traverse, /*tp_traverse*/
     (inquiry)WraptObjectProxy_clear, /*tp_clear*/
     (richcmpfunc)WraptObjectProxy_richcompare, /*tp_richcompare*/
-    0,                      /*tp_weaklistoffset*/
+    offsetof(WraptObjectProxyObject, weakreflist), /*tp_weaklistoffset*/
     (getiterfunc)WraptObjectProxy_iter, /*tp_iter*/
     0,                      /*tp_iternext*/
     WraptObjectProxy_methods, /*tp_methods*/
@@ -1754,7 +1759,7 @@ PyTypeObject WraptCallableObjectProxy_Type = {
     0,                      /*tp_traverse*/
     0,                      /*tp_clear*/
     0,                      /*tp_richcompare*/
-    0,                      /*tp_weaklistoffset*/
+    offsetof(WraptObjectProxyObject, weakreflist), /*tp_weaklistoffset*/
     0,                      /*tp_iter*/
     0,                      /*tp_iternext*/
     0,                      /*tp_methods*/
@@ -2249,7 +2254,7 @@ PyTypeObject WraptFunctionWrapperBase_Type = {
     (traverseproc)WraptFunctionWrapperBase_traverse, /*tp_traverse*/
     (inquiry)WraptFunctionWrapperBase_clear, /*tp_clear*/
     0,                      /*tp_richcompare*/
-    0,                      /*tp_weaklistoffset*/
+    offsetof(WraptObjectProxyObject, weakreflist), /*tp_weaklistoffset*/
     0,                      /*tp_iter*/
     0,                      /*tp_iternext*/
     0,                      /*tp_methods*/
@@ -2482,7 +2487,7 @@ PyTypeObject WraptBoundFunctionWrapper_Type = {
     0,                      /*tp_traverse*/
     0,                      /*tp_clear*/
     0,                      /*tp_richcompare*/
-    0,                      /*tp_weaklistoffset*/
+    offsetof(WraptObjectProxyObject, weakreflist), /*tp_weaklistoffset*/
     0,                      /*tp_iter*/
     0,                      /*tp_iternext*/
     0,                      /*tp_methods*/
@@ -2622,7 +2627,7 @@ PyTypeObject WraptFunctionWrapper_Type = {
     0,                      /*tp_traverse*/
     0,                      /*tp_clear*/
     0,                      /*tp_richcompare*/
-    0,                      /*tp_weaklistoffset*/
+    offsetof(WraptObjectProxyObject, weakreflist), /*tp_weaklistoffset*/
     0,                      /*tp_iter*/
     0,                      /*tp_iternext*/
     0,                      /*tp_methods*/